// apps/frontend/src/types/schema.ts
export interface Database {
// ...
  public: {
    Tables: {
// ...
      pachislo_payments: {
        Row: {
          created_at: string | null
          id: number
          kind: string
          machine: number
          rate: number
          shop: number
          update_at: string | null
        }
        Insert: {
          created_at?: string | null
          id?: number
          kind: string
          machine: number
          rate: number
          shop: number
          update_at?: string | null
        }
        Update: {
          created_at?: string | null
          id?: number
          kind?: string
          machine?: number
          rate?: number
          shop?: number
          update_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "pachislo_payments_machine_fkey"
            columns: ["machine"]
            referencedRelation: "machine"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "pachislo_payments_rate_fkey"
            columns: ["rate"]
            referencedRelation: "rate"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "pachislo_payments_shop_fkey"
            columns: ["shop"]
            referencedRelation: "shop"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}